xmlstream: support readers that return nil tokens
While working on a feature in mellium.im/xmpp, I noticed a deadlock when
copying tokens from an xml.Decoder.  After stepping through the code,
the issue appeared to be that xml.Decoder was returning nil, nil and
that xmlstream.Copy was entering an infinite loop copying the nil tokens
into a noop encoder (which does not verify the correctness of tokens,
which would have resulted in an error on a nil token).  Cloning Go 1.14
and opening encoding/xml/xml.go around line 290 I found a check for
io.EOF that reset the error to nil that was introduced by past-me
(oops!).

Since this behavior on Decoders can't be changed quickly, several tests
were added to double check that anything that iterates over the token
stream terminates when nil, nil is encountered as well as nil, io.EOF.
In future, it may be possible to change xml.Decoder's to return nil,
io.EOF any time nil, nil would be returned.

The encoding/xml package is a hot mess, and it's at least partially my
fault. I'm sorry.

Signed-off-by: Sam Whited <sam@samwhited.com>