.builds: don't use -cover or tests will fail
This is a fun one that amuses me every time I see it. TL;DR we have a
test that makes sure we don't do expensive memory allocations in some
code that needs to be very efficient. Unfortunately, using -cover
instruments our code so that it can tell which branches have been called
and which ones haven't and the code it injects performs mallocs, causing
the test to fail.

Signed-off-by: Sam Whited <sam@samwhited.com>