xmlstream: fix early EOF on InnerElement
Previously if the end element to be returned was also the last element
in the stream and was returned at the same time as an early EOF error it
would be incorrectly skipped.
Removing the unnecessary error checking after popping the token fixes
this by not short circuiting when it sees EOF.
Errors are still returned at the end of the function, but only after any
start/end elements have been handled.

Signed-off-by: Sam Whited <sam@samwhited.com>